+++ joulescope_ui/main.py
@@ -35,15 +35,17 @@ from joulescope.data_recorder import construct_record_filename
 from joulescope_ui.recording_viewer_device import RecordingViewerDevice
 from joulescope_ui.preferences import PreferencesDialog
 from joulescope_ui.config import load_config_def, load_config, save_config
-from joulescope_ui import usb_inrush
 from joulescope_ui.update_check import check as software_update_check
 from joulescope_ui.logging_util import logging_config
 from joulescope_ui.oscilloscope.signal_statistics import si_format, html_format, three_sig_figs
-from joulescope_ui.exporter import Exporter
+from joulescope_ui.range_tool import RangeToolInvoke
 from joulescope_ui import help_ui
 from joulescope_ui import firmware_manager
+from joulescope_ui.plugin_manager import PluginManager
+from joulescope_ui.exporter import Exporter
 import io
 import ctypes
+import collections
 import traceback
 import webbrowser
 import logging
@@ -102,6 +104,19 @@ class ValueLabel(QtWidgets.QLabel):
         self.setText('%s=%-8.2f' % (self._text, value))
 
 
+def dict_update_recursive(tgt, src):
+    """Merge src into tgt, overwriting as needed.
+
+    :param tgt: The target dict.
+    :param src: the source dict.
+    """
+    for key, value in src.items():
+        if key in tgt and isinstance(tgt[key], dict) and isinstance(value, collections.Mapping):
+            dict_update_recursive(tgt[key], value)
+        else:
+            tgt[key] = value
+
+
 class DeviceDisable:
     def __init__(self):
         self.view = NullView()
@@ -158,8 +173,9 @@ class MainWindow(QtWidgets.QMainWindow):
         if cfg is None:
             self._cfg = load_config(self._cfg_def, default_on_error=True)
         else:
-            self._cfg = cfg
+            self._cfg = cfg  # warning: shared mutable data, update only (no replace)
         self._path = self._cfg['General']['data_path']
+        self._plugins = PluginManager(self._cfg)
 
         super(MainWindow, self).__init__()
         self.ui = Ui_mainWindow()
@@ -234,7 +250,7 @@ class MainWindow(QtWidgets.QMainWindow):
 
         # Oscilloscope: current, voltage, power, GPI0, GPI1, i_range
         self.oscilloscope_dock_widget = QtWidgets.QDockWidget('Waveforms', self)
-        self.oscilloscope_widget = Oscilloscope(self.oscilloscope_dock_widget)
+        self.oscilloscope_widget = Oscilloscope(self.oscilloscope_dock_widget, plugins=self._plugins)
         self.oscilloscope_dock_widget.setVisible(False)
         self.oscilloscope_dock_widget.setWidget(self.oscilloscope_widget)
         self.addDockWidget(QtCore.Qt.RightDockWidgetArea, self.oscilloscope_dock_widget)
@@ -282,7 +298,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.oscilloscope_widget.sigMarkerDualAddRequest.connect(self.on_markerDualAddRequest)
         self.oscilloscope_widget.sigMarkerRemoveRequest.connect(self.on_markerRemoveRequest)
         self.oscilloscope_widget.sigMarkerDualUpdateRequest.connect(self.on_markerDualUpdateRequest)
-        self.oscilloscope_widget.sigExportDataRequest.connect(self.on_exportData)
+        self.oscilloscope_widget.sigRangeToolRequest.connect(self.on_rangeTool)
 
         # status update timer
         self.status_update_timer = QtCore.QTimer(self)
@@ -317,8 +333,9 @@ class MainWindow(QtWidgets.QMainWindow):
 
         # tools
         self.ui.actionClearEnergy.triggered.connect(self._tool_clear_energy)
-        self.ui.actionUsbInrush.triggered.connect(self._tool_usb_inrush)
-        self.ui.actionUsbInrush.setEnabled(usb_inrush.is_available())
+        with self._plugins as p:
+            p.range_tool_register('Export data', Exporter)
+        self._plugins.builtin_register(app_config=self._cfg)
 
         self._dock_widgets = [
             self.dev_dock_widget,
@@ -328,9 +345,6 @@ class MainWindow(QtWidgets.QMainWindow):
             self.uart_dock_widget,
             self.oscilloscope_dock_widget,
         ]
-
-        self._exporter = Exporter(self)
-        self._exporter.sigFinished.connect(self.on_exporterFinished)
 
         self._device_close()
         self._cfg_apply()
@@ -528,11 +542,6 @@ class MainWindow(QtWidgets.QMainWindow):
         log.info('_tool_clear_energy: offset= %g J, offset=%g C', self._energy[0], self._charge[0])
         self._charge[1] = self._charge[0]
         self._energy[1] = self._energy[0]
-
-    def _tool_usb_inrush(self):
-        data = self._device.view.extract()
-        rv = usb_inrush.run(data, self._device.view.sampling_frequency)
-        self.status(f'USB inrush analysis completed: return code {rv}')
 
     def _source_indicator_set(self, text, color=None, tooltip=None):
         tooltip = '' if tooltip is None else str(tooltip)
@@ -1214,7 +1223,7 @@ class MainWindow(QtWidgets.QMainWindow):
             cfg['Device']['i_range_update'] = (self._cfg['Device']['i_range'] != cfg['Device']['i_range'])
             cfg['Device']['v_range_update'] = (self._cfg['Device']['v_range'] != cfg['Device']['v_range'])
 
-            self._cfg = cfg
+            dict_update_recursive(self._cfg, cfg)
             save_config(self._cfg)
             self._cfg_apply()
 
@@ -1260,17 +1269,42 @@ class MainWindow(QtWidgets.QMainWindow):
             html = html_format(txt_result)
             m2.html_set(key, html)
 
-    @QtCore.Slot(float, float)
-    def on_exportData(self, x_start, x_stop):
-        self._exporter.export(self._device.view, x_start, x_stop, path_default=self._path)
+    @QtCore.Slot(str, float, float)
+    def on_rangeTool(self, name, x_start, x_stop):
+        range_tool = self._plugins.range_tools.get(name)
+        if range_tool is None:
+            self.status('Range tool not found')
+            return
+        plugin_config = self._cfg.get('plugins', {}).get(name, {})
+        invoke = RangeToolInvoke(self, range_tool, app_config=self._cfg, plugin_config=plugin_config)
+        invoke.sigFinished.connect(self.on_rangeToolFinished)
+        s = self._device.statistics_get(x_start, x_stop)
+        invoke.run(self._device.view, s, x_start, x_stop)
 
-    @QtCore.Slot(str)
-    def on_exporterFinished(self, msg):
+    @QtCore.Slot(object, str)
+    def on_rangeToolFinished(self, range_tool, msg):
         if msg:
             log.warning(msg)
             self.status(msg)
         else:
-            self.status('Export done')
+            self.status(range_tool.name + ' done')
+
+    def range_tool_menu_construct(self, menu=None):
+        instances = []  # hold on to QT objects
+        if menu is None:
+            menu = QtGui.QMenu()
+            menu.setToolTipsVisible(True)
+        export_data = QtGui.QAction('&Export data', self)
+        export_data.triggered.connect(self._range_tool_constructor('export'))
+        menu.addAction(export_data)
+        tools = self._axis().range_tools
+        for name, in tools.keys():
+            t = QtGui.QAction(name, self)
+            t.triggered.connect(self._analysis_menu_callback_constructor(name))
+            menu.addAction(t)
+            instances.append(t)
+        menu.instances = instances
+        return menu
 
 
 class ErrorWindow(QtWidgets.QMainWindow):